Extract table name from  multiline select statements
Move away from using `$1` global as it worries me that we may have many
thousands of requests being handled by a small number of threads, thus
contending on a global variable. No idea if it'll be a problem, but
would rather avoid it.

Use a lazy enumeration to map/detect from the most common query type to
least common (select -> insert -> update -> delete).