fix incorrect sidekiq metric names
The metric name for Sidekiq workers currently have an empty label in
them, e.g. `ActivityPub..DeliveryWorker`. This is because of the use of
`tr`, which translates character by character, and is not a substring
substitution. Using `gsub` fixes the problem.

Unfortunately, this does mean that dashboards are probably going to
break everywhere with this change.