Test all floats within delta
Float math isn't exact, so it's better to assert that the value falls
within a small distance of the desired value.

Running this suite I see deltas of 0.0002 so adding a delta of 0.001
should pass consistently and still check for extremely accurate results
(we are looking at a value in milliseconds).